refactor: model 부분 controller으로 통합
