--- src/components/main/heroSection/heroSectionModel.ts
@@ -1,11 +1,0 @@
-export const intervalIdState = {
-  intervalId: 0,
-
-  setIntervalId(id: number) {
-    this.intervalId = id;
-  },
-
-  getIntervalId() {
-    return this.intervalId;
-  },
-};
